requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord==1.0.1
discord.py==1.7.2
googlesearch-python==2020.0.2
idna==2.10
multidict==5.1.0
requests==2.24.0
soupsieve==2.0.1
typing-extensions==3.10.0.0
urllib3==1.25.9
yarl==1.6.3
